feat: add backend streaming...

+++ backend/stream/services/stream.py
@@ -17,10 +17,14 @@ class Stream:
         response = requests.get(self.torrent_url)
 
         if response.status_code == 200:
-            with open(f"/tmp/torrent_files/{self.movie_name}.torrent", 'wb') as f:
-                f.write(response.content)
-            print("Torrent file downloaded successfully.")
-            self.torrent_file_path = f"/tmp/torrent_files/{self.movie_name}.torrent"
+            file_path = f"/tmp/torrent_files/{self.movie_name}.torrent"
+            if not os.path.exists(file_path):
+                with open(f"/tmp/torrent_files/{self.movie_name}.torrent", 'wb') as f:
+                    f.write(response.content)
+                print("Torrent file downloaded successfully.")
+            else:
+                print("Torrent file already exists.")
+            self.torrent_file_path = file_path
 
         else:
             print(
@@ -47,52 +51,138 @@ class Stream:
         print(f"Downloaded file is saved in: {save_path}")
         
     def stream_torrent(self):
-        ses = lt.session()
-        ses.listen_on(6881, 6891)
-        torrent_info = lt.torrent_info(self.torrent_file_path)
-        h = ses.add_torrent({'ti': torrent_info, 'save_path': './'})
+        if not self.torrent_file_path:
+            return HttpResponse("Torrent file path is required!", status=400)
         
-        print("Downloading metadata...")
-        while not h.has_metadata():
-            time.sleep(1)
-        print("Metadata received, starting download...")
-        torrent_info = h.get_torrent_info()
-        while not h.have_piece(1):
-            time.sleep(0.1)
-        print("test ===> ", h.have_piece(1))
-        def get_video_chunk(start_byte, chunk_size=1024*1024):
-            """Function to get a chunk of video from the torrent."""
-            # Download the necessary chunk
-            piece_start = start_byte
-            piece_end = start_byte + chunk_size
-            torrent_info = h.get_torrent_info()
-            print("torrent_info: ", torrent_info)
-            # pieces = torrent_info.files()
+        # Create a session
+        session = lt.session()
+        session.listen_on(6881, 6891)
+        
+        
+        # Add a torrent
+        print(f"====> Adding torrent from: {self.torrent_file_path}")
+        torrent_info = lt.torrent_info(self.torrent_file_path)
+        handle = session.add_torrent({'ti': torrent_info, 'save_path': './'})
+        
+        
+        # get the info hash
+        selected_file = torrent_info.files()
+        print(f"====> Selected file: {selected_file}")
+        file_size = selected_file.piece_size(0)
+        file_name = selected_file.file_name(0)
+        video_path = os.path.join('./', selected_file.file_path(0))
+        num_of_pieces = torrent_info.num_pieces()
+        print(f"====> Selected file: {file_name} - {file_size} - {num_of_pieces} - video_path: {video_path}")
 
-            # file = pieces[0]
-            # file_size = file.size
 
-            # # Check for the end of the file
-            # if piece_start >= file_size:
-            #     return None
+        # Wait for the metadata
+        print("====> Waiting for metadata...")
+        while not handle.status().has_metadata:
+            time.sleep(.5)
+            
+        # Start streaming
+        print("====> reading on peace...")
+        # for piece_index in range(torrent_info.num_pieces()):
+            
+        handle.read_piece(0)
+        while not handle.have_piece(0):
+            time.sleep(.1)
+        session.remove_torrent(handle)
 
-            # # Read the next chunk
-            # with open(file.path, 'rb') as f:
-            #     f.seek(piece_start)
-            #     chunk = f.read(min(chunk_size, file_size - piece_start))
+        with open(video_path, 'rb') as f:
+            f.seek(0)
+            print("chunk:", f.read(100))
+        
 
-            return "ff"
 
-        def video_stream():
-            start_byte = 0
-            while start_byte < os.path.getsize(self.torrent_file_path):
-                chunk = get_video_chunk(start_byte)
-                if chunk is None:
-                    break
-                start_byte += len(chunk)
-                yield chunk
 
-        response = StreamingHttpResponse(video_stream(), content_type='video/mp4')
-        response['Content-Disposition'] = 'inline; filename="video.mp4"'
+        return HttpResponse("Streaming completed successfully.")
+
+
+SAVE_PATH = './_Movies'
+TORRENT_FILES_PATH = '/tmp/torrent_files'
+
+
+class TorrentStream:
+    def __init__(self):
+        if not os.path.exists("/tmp/torrent_files"):
+            os.makedirs("/tmp/torrent_files")
+        self.torrent_file_path = None
+        self.session = lt.session()
+        self.torrent_info = None
+        self.handle = None
+        self.session.listen_on(6881, 6891)
+
+        self.movie_path = None
+        self.file_size = None
+        self.start_byte = 0
+        self.end_byte = None
+        self.piece_size = None
+
+    def init_torrent_file(self, torrent_url):
+        response = requests.get(torrent_url)
+        self.torrent_file_path = f"{TORRENT_FILES_PATH}/{torrent_url.split('/')[-1].split('.')[0]}.torrent"
+
+        if response.status_code == 200:
+            if not os.path.exists(self.torrent_file_path):
+                with open(self.torrent_file_path, 'wb') as f:
+                    f.write(response.content)
+                print("Torrent file downloaded successfully.")
+            else:
+                print("Torrent file already exists.")
+        else:
+            print(
+                f"Failed to download the torrent. Status code: {response.status_code}")
+
+    def add_torrent(self):
+        print(f"====> Adding torrent from: {self.torrent_file_path}")
+        self.torrent_info = lt.torrent_info(self.torrent_file_path)
+        self.handle = self.session.add_torrent({'ti': self.torrent_info, 'save_path': SAVE_PATH, 'storage_mode': lt.storage_mode_t.storage_mode_sparse})
+
+        print("====> Waiting for metadata...")
+        while not self.handle.status().has_metadata:
+            time.sleep(.5)
+        print("====> Metadata received successfully.")
+
+        self.movie_path = os.path.join(SAVE_PATH, self.torrent_info.files().file_path(0))
+        self.file_size = self.torrent_info.files().file_size(0)
+
+    def parse_chunk_range(self, vrange):
+        start, end = vrange.replace('bytes=', '').split('-')
+        start = int(start)
+        end = int(end) if end else None
+        self.piece_size = self.torrent_info.piece_length()
+        print(f"====> Piece size: {self.piece_size}")
+        piece_index = start // self.piece_size
+        return {'piece_index': piece_index, 'start': start, 'end': end}
+
+    def stream_torrent(self, vrange):
+        parsed_range = self.parse_chunk_range(vrange)
+        piece_index = parsed_range['piece_index']
+        self.start_byte = parsed_range['start']
+        self.end_byte = parsed_range['end']
+
+        print(f"====> Reading piece: {piece_index}")
+        self.handle.read_piece(piece_index)
+        while not self.handle.have_piece(piece_index):
+            time.sleep(.1)
+        
+        
+
+        video_path = os.path.join(SAVE_PATH, self.torrent_info.files().file_path(0))
+
+        with open(video_path, 'rb') as f:
+            f.seek(0)
+            return f.read(self.torrent_info.piece_length())
+
+    def create_response(self, data):
+        response = StreamingHttpResponse(data, content_type="video/mp4", status=206)
+        self.end_byte = self.start_byte + len(data) - 1
+        print(f"====> Start byte: {self.start_byte} - End byte: {self.end_byte} - File size: {self.file_size}")
+        response['Content-Range'] = f'bytes {self.start_byte}-{self.end_byte}/{self.file_size}'
+        response['Content-Length'] = len(data)
+        response['Accept-Ranges'] = 'bytes'
         return response
-        # return {'status': 'success', 'message': 'Stream torrent successfully!'}
+
+    def remove_stream(self):
+        self.session.remove_torrent(self.handle)
